Bug fix for copy to clipboard button
Test: go/vts-web-staging
Bug: 110814448

+++ src/main/java/com/android/vts/entity/CoverageEntity.java
@@ -268,8 +268,8 @@ public class CoverageEntity implements Serializable {
     List<Long> lineCoverage = null;
     if (coverage.getLineCoverageVectorCount() > 0) {
       lineCoverage = new ArrayList<>();
-      for (int count : coverage.getLineCoverageVectorList()) {
-        lineCoverage.add((long) count);
+      for (Long count : coverage.getLineCoverageVectorList()) {
+        lineCoverage.add(count);
       }
     }
     return new CoverageEntity(parentKey, group, coveredLineCount, totalLineCount, filePath,
